Update main.py
As you might notice, there is a drastic difference in this version of the main.py file from the original branch. This reflects a change in approach on this project. When I first started this project, the approach was a very rudimentary one...

For instance, the method for retrieving chat data was a beautiful soup & selenium based scraper that scraped the front-end.