Create channel_scraper.py
This was built as an attempt to get the chat room id of the selected streamer (emphasis on attempt). The actual functionality to get the channel id autonomously is not yet working...however I am confident that it will be fixed in the future. 

In the meantime, I made it so that if you have the chatroom id of the streamer, you can use it. Unfortunately, it can be an involved process for people not accustomed with using browser developer tools so I will include the instructional on doing that in the near future.

However this is the basic manner of attaining the chat room id: 
1. press f12
2. Go to Network/Network Traffic (may look like a little WI-FI icon)
3. Select WS. It will be empty at first.
4. Refresh. This will populate with one websocket connection that the website makes when connecting to the chatroom. 
5. Click on the connection.
6. Select any of the json entries beginning with "event". 
7. Copy the id value in between the chatrooms. and v2. line. For instance: chatrooms.135413.v2". In this case the chatroom id will be 135413.
8. That's it! Thats the chatroom id you will be prompted to enter when the autonomous scraper portion of this module indoubtedly fails lol. For now, I will try to manually add as many chatroom id entries until I can get this autonomous chatroom id scraper working.
Thanks for reading!